(WIP) example analysis: pvl-delta on IGT data
Autoreversediff in combination with the current implementation of
updating expected values by mutating a vector blows up the ram.
This happens when estimating a parameter for learning rate, but not when
only estimating for temperature.
ForwardDiff doesn't seem to blow up ram, but it's incredibly slow.
Other Autodifferentiation packages yet untested.